third_party/WebKit/LayoutTests/svg/hixie/error/001-expected.txt

CONSOLE ERROR: line 3: Error: <circle> attribute r: Expected length, "r".
